api/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
python-multipart==0.0.6
pydantic==2.5.0
redis==5.0.1
requests==2.31.0
python-jose==3.3.0
passlib==1.7.4
python-dateutil==2.8.2
pytz==2023.3
phonenumbers==8.13.27
python-dotenv==1.0.1
geopy==2.4.1